http: Generalize RTSP parsing to HTTP
Move parse_rtsp from raop to a new http module under support and make it
more general, to support any protocol with the same format. As AirPlay
multiplexes both HTTP and RTSP over the same connection, this parser
will be used to support that.

Relates to #1078
